auxilium_planner/requirements.txt
# FastAPI and server
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.11.0
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Additional dependencies
python-multipart==0.0.9
aiofiles==23.2.1

# Development and testing
pytest==8.1.1
pytest-asyncio==0.23.5
httpx==0.28.0

# Utilities
uuid==1.30
datetime==5.4 